refactor: extract `just<class,class>` into header
